requirements.txt
numpy==1.22.4
Pillow==9.1.1
streamlit==1.10.0
tensorflow==2.9.1
